#124 [refactor] patchBook, postCreateBook usecase 적용
