+++ app/src/main/java/com/sopt/peekabookaos/domain/usecase/PatchBookUseCase.kt
@@ -1,0 +1,11 @@
+package com.sopt.peekabookaos.domain.usecase
+
+import com.sopt.peekabookaos.domain.repository.CreateUpdateRepository
+import javax.inject.Inject
+
+class PatchBookUseCase @Inject constructor(
+    private val createUpdateRepository: CreateUpdateRepository
+) {
+    suspend operator fun invoke(bookId: Int, description: String?, memo: String?) =
+        createUpdateRepository.patchBook(bookId, description, memo)
+}
